requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
argcomplete==2.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
Beaker==1.12.1
beautifulsoup4==4.12.2
blivet==3.8.2
blivet-gui==2.4.2
Brlapi==0.8.5
btrfs-backup==0.3.1
certifi==2023.11.17
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
cryptography==41.0.3
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.7
distro==1.8.0
dnf==4.18.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.13.1
fros==1.1
frozenlist==1.4.0
gpg==1.20.0
gql==3.4.1
graphql-core==3.2.3
humanize==3.13.1
idna==3.6
iniconfig==1.1.1
langtable==0.0.64
libcomps==0.1.20
libdnf==0.72.0
lxml==4.9.3
Mako==1.2.3
MarkupSafe==2.1.3
multidict==6.0.4
nftables==0.1
olefile==0.46
packaging==23.1
Paste==3.5.3
pbr==6.0.0
pexpect==4.8.0
pid==2.2.3
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.2.0
ply==3.11
powerline-status==2.8.3
productmd==1.37
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pygit2==1.13.3
PyGObject==3.46.0
pykickstart==3.48
pyOpenSSL==23.2.0
pyparted==3.13.0
PySocks==1.7.1
pytest==7.3.2
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-meh==0.51
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rpm==4.19.0
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.6.0
soupsieve==2.5
speedtest-cli==2.1.3
stevedore==5.1.0
systemd-python==235
Tempita==0.5.2
urllib3==2.1.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
yarl==1.9.4
